Mobile Navbar Collision fix
HeaderStyle.css:
Fixed the animation positioning for the second collapsing menu. Only one issue to fix in the future. If a user spams the two buttons they may experience a little navbar collision since multiple calls are running faster than it can animate. For the sake of time it will be done for now but easiest solution would be a redesign.

header.js:
Refactored the code again. Created a function constructor to copy behaviours and get updates from each object individually to avoid modal collisions. Also, created a window event to take care of if we click outside our modal window.