Page Data Logic
Issue: Data was not displaying correctly on the console using the getPageData method.

Fix: The issue was that the startIndex was off by one because I was adding a +1 to my start index. This is due to a misunderstanding of how slice indexes work. I tested many possible outcomes, all outcomes came out correct now.

Next: Change the table data.